// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      occupancy: {
        Row: {
          check_in_date: string
          check_out_date: string | null
          created_at: string | null
          id: string
          is_current: boolean | null
          latest_meter_reading: number | null
          room_id: string
          tenant_id: string
          updated_at: string | null
        }
        Insert: {
          check_in_date: string
          check_out_date?: string | null
          created_at?: string | null
          id?: string
          is_current?: boolean | null
          latest_meter_reading?: number | null
          room_id: string
          tenant_id: string
          updated_at?: string | null
        }
        Update: {
          check_in_date?: string
          check_out_date?: string | null
          created_at?: string | null
          id?: string
          is_current?: boolean | null
          latest_meter_reading?: number | null
          room_id?: string
          tenant_id?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "fk_occupancy_room"
            columns: ["room_id"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "fk_occupancy_tenant"
            columns: ["tenant_id"]
            isOneToOne: false
            referencedRelation: "tenants"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "occupancy_room_id_fkey"
            columns: ["room_id"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "occupancy_tenant_id_fkey"
            columns: ["tenant_id"]
            isOneToOne: false
            referencedRelation: "tenants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      rooms: {
        Row: {
          capacity: number
          created_at: string | null
          floor: number
          id: string
          price: number
          room_number: string
          room_type: string
          status: string
          tenant_id: string | null
          updated_at: string | null
        }
        Insert: {
          capacity: number
          created_at?: string | null
          floor: number
          id?: string
          price: number
          room_number: string
          room_type: string
          status?: string
          tenant_id?: string | null
          updated_at?: string | null
        }
        Update: {
          capacity?: number
          created_at?: string | null
          floor?: number
          id?: string
          price?: number
          room_number?: string
          room_type?: string
          status?: string
          tenant_id?: string | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "rooms_tenants_fkey"
            columns: ["tenants"]
            isOneToOne: false
            referencedRelation: "tenants"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      tenants: {
        Row: {
          address: string | null
          auth_email: string | null
          created_at: string | null
          email: string | null
          emergency_contact: string | null
          first_name: string
          id: string
          last_name: string
          phone: string | null
          room_id: string
          room_number: string
          updated_at: string | null
          residents: string
        }
        Insert: {
          address?: string | null
          auth_email?: string | null
          created_at?: string | null
          email?: string | null
          emergency_contact?: string | null
          first_name: string
          id?: string
          last_name: string
          phone?: string | null
           room_id: string
          room_number: string
          updated_at?: string | null
          residents: string
        }
        Update: {
          address?: string | null
          auth_email?: string | null
          created_at?: string | null
          email?: string | null
          emergency_contact?: string | null
          first_name?: string
          id?: string
          last_name?: string
          phone?: string | null
          room_id: string
          room_number?: string
          updated_at?: string | null
          residents: string
        }
        Relationships: [
          {
            foreignKeyName: "tenants_room_number_fkey"
            columns: ["room_number"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["room_number"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      is_admin: {
        Args: Record<PropertyKey, never>
        Returns: boolean
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}